refactor: 코드 리뷰 피드백 반영
like type 파라미터 union타입 사용
req.user 값에 모든 데이터를 넣지 않고 id, email, nickname만 넣기
중복된 데코레이터 제거
상태코드 204 -> 200
strict equal로 변경

+++ src/api/controllers/article.controller.ts
@@ -10,6 +10,11 @@ import { FindManyParamsDto } from "../types/dto.js";
 const ArticleController = {
   async createArticle(req: RequestWithDto<ArticleDto>, res: Response, next: NextFunction) {
     try {
+      if (!req.user) {
+        const error: CustomError = new Error("인증이 필요합니다.");
+        error.statusCode = 401;
+        throw error;
+      }
       const { id: userId } = req.user;
       const articleDto = req.body;
       const newArticle = await ArticleService.createArticle(articleDto, userId);
@@ -50,6 +55,11 @@ const ArticleController = {
 
   async updateArticle(req: RequestWithDto<ArticleDto>, res: Response, next: NextFunction) {
     try {
+      if (!req.user) {
+        const error: CustomError = new Error("인증이 필요합니다.");
+        error.statusCode = 401;
+        throw error;
+      }
       const { id: userId } = req.user;
       const { id } = req.params;
       const articleDto = req.body;
@@ -63,10 +73,15 @@ const ArticleController = {
 
   async deleteArticle(req: Request, res: Response, next: NextFunction) {
     try {
+      if (!req.user) {
+        const error: CustomError = new Error("인증이 필요합니다.");
+        error.statusCode = 401;
+        throw error;
+      }
       const { id: userId } = req.user;
       const { id } = req.params;
       await ArticleService.deleteArticle(Number(id), userId);
-      res.status(204).json({ success: "상품 삭제 성공" });
+      res.status(200).json({ success: "상품 삭제 성공" });
     } catch (err) {
       next(err);
     }
